add warp bit transpose

+++ include/cuda/Common.h
@@ -31,7 +31,7 @@ using DIM = uint32_t;
 using QUANTIZED_INT = int;
 using SERIALIZED_TYPE = unsigned char;
 using Byte = unsigned char;
-using OPTION = uint32_t;
+using OPTION = int8_t;
 }
 
 #include <iostream>
